Point a mobile's position at the Position type that was already there
Step 2's last enumeration, and its smallest: `MobDef.Position` and
`MobDef.DefaultPosition` stop being `int32` and become `game.Position`.

§3.2 calls this one out by name -- "note in particular that
`game.Position` is a defined type and `MobDef.Position` is an `int32`
holding the same domain". Nothing had to be designed; the type, its nine
constants, its `String()` and its `Awake()` have been in `position.go`
since the port. The prototype's two fields were simply never pointed at
it, so the same domain had two representations depending on whether you
were reading a live character or the file it was loaded from.

Two conversions disappear rather than move, which is the useful sign
that the field was always this type:

    reset.go:273   Position: Position(def.Position)   -> def.Position
    commands.go:1354
      who.Position == game.Position(who.MobDef.Position)
                                            -> who.MobDef.Position

That second one is `listOneChar` deciding whether a mobile is standing
as its long description describes it. It was comparing a `Position`
against an `int32` *through a conversion the author had to write*, which
is the shape §3.2 is about: the compiler asked for the conversion, got
it, and had nothing left to check.

Four boundaries convert instead, all of them writing the number to a
file or a name table: `world/dump.go`'s JSON, `world/yaml`'s reader and
writer, and `world/classic`'s mob reader. Those are explicit `int32(...)`
rather than a `Number()` accessor, because `Position` is already `int32`
underneath -- unlike the enumerations declared in this step, which are
`int`, and which need `Number()` to hold their one G115 suppression.
Whether `Position` should become `int` too is §3.8's question and stays
there.

Position has a `String()` method, so `%v` on one now prints "standing"
rather than "8". Every site that writes the number does so through an
explicit conversion, so nothing changed -- but that is the hazard to
watch for on this type specifically, and it is why `make parity`
matters here more than for the other enumerations: a mobile's position
round-trips through both world readers and the dump the C is compared
against.

**Step 2 is done.** Nine domains: Class, Sex, ItemType, Apply, Sector,
Race, SpellID, Liquid, Position. The plan's status block records what
four of them turned out not to be.

Checks:

    make lint         0 issues
    make test-fast    green
    make play         green (147s, -race)
    make parity       identical: 1878 rooms, 569 mobiles, 679 objects,
                      30 zones, 46 shops

+++ internal/game/reset_test.go
@@ -26,9 +26,9 @@ func resetWorld(commands []ResetCommand) *Live {
 				LongDesc: "A cityguard stands here.\r\n",
 				Level:    10, Thac0: 15, ArmorClass: 5,
 				HitDice: Dice{Number: 2, Size: 8, Bonus: 10},
-				Gold:    100, Exp: 500, Position: int32(PosStanding),
+				Gold:    100, Exp: 500, Position: PosStanding,
 			},
-			{Vnum: 3061, Keywords: "dog", ShortDesc: "a dog", Level: 2, Position: int32(PosStanding)},
+			{Vnum: 3061, Keywords: "dog", ShortDesc: "a dog", Level: 2, Position: PosStanding},
 		},
 		Objects: []*ObjDef{
 			{Vnum: 3020, Keywords: "sword", ShortDesc: "a sword", Type: ItemWeapon,
@@ -349,7 +349,7 @@ func freshGuardDef() *MobDef {
 		LongDesc: "A reloaded cityguard stands here.\r\n",
 		Level:    20, Thac0: 5, ArmorClass: 1,
 		HitDice: Dice{Number: 4, Size: 8, Bonus: 20},
-		Gold:    999, Exp: 5000, Position: int32(PosStanding),
+		Gold:    999, Exp: 5000, Position: PosStanding,
 		ActionFlags: NewSet(MobIsNPC, MobAggrNeutral), // an arbitrary extra flag, not just the always-set one
 	}
 }
@@ -534,9 +534,9 @@ func shopWorld() *Live {
 		Rooms: []*RoomDef{temple},
 		Mobiles: []*MobDef{
 			{Vnum: 3060, Keywords: "keeper", ShortDesc: "the keeper",
-				Level: 10, Position: int32(PosStanding)},
+				Level: 10, Position: PosStanding},
 			{Vnum: 3061, Keywords: "second keeper", ShortDesc: "the second keeper",
-				Level: 10, Position: int32(PosStanding)},
+				Level: 10, Position: PosStanding},
 		},
 		Shops: []*ShopDef{
 			{Vnum: 9001, Keeper: 3060, ProfitBuy: 1.1, ProfitSell: 0.5, Temper: 0},
